+++ apps/web/src/methods/index.ts
@@ -4,8 +4,10 @@ import { getAttestation } from './server/getAttestation'
 import { getCollegeInfo } from './server/getCollegeInfo'
 import { getLessons } from './server/getLessons'
 import { getPerformance } from './server/getPerformance'
+import { getFinalMarks } from './server/getFinalMarks'
 
 export {
+  getFinalMarks,
   getLessons,
   getPerformance,
   getAds,
