+++ apps/web/src/types/index.ts
@@ -39,3 +39,5 @@ export type ReturnedMark =
   | 'ДЗ'
   | '.'
   | number
+
+export type Nullable<T> = T | null
