+++ apps/web/src/views/Attestation/FinalMarks/index.tsx
@@ -1,7 +1,7 @@
 import { AcademicRecord } from '@diary-spo/shared'
 import { FC } from 'preact/compat'
 import { Table } from './Table'
-import { buildSubjectMatrix } from './Table/buildSubjectMatrix.ts'
+
 import { ScreenSpinner } from '@vkontakte/vkui'
 
 import './index.css'
@@ -20,11 +20,9 @@ const FinalMarks: FC<Props> = ({ isDataLoading, data }) => {
     return
   }
 
-  const subjectMatrix = buildSubjectMatrix(data)
-
   return (
     <div className='tableWrapper'>
-      <Table subjectMatrix={subjectMatrix} />
+      <Table data={data} />
     </div>
   )
 }
