+++ apps/web/tsconfig.json
@@ -31,6 +31,7 @@
       "@store": ["./src/store"],
       "@libs": ["./src/libs"],
       "@hooks": ["./src/hooks"],
+      "@types": ["./src/types"],
     }
   },
   "include": ["src"],
